- switched to .bz2, cleanups, use macro for texmf dir, fixed preun/postun
Changed files:
    lyx.spec -> 1.74

+++ lyx.spec
@@ -7,8 +7,8 @@ Release:	1
 Epoch:		1
 License:	GPL
 Group:		Applications/Publishing/TeX
-Source0:	ftp://ftp.lyx.org/pub/lyx/stable/%{name}-%{version}.tar.gz
-# Source0-md5:	9be5ddcc9b9477f7aa5342974c21195b
+Source0:	ftp://ftp.lyx.org/pub/lyx/stable/%{name}-%{version}.tar.bz2
+# Source0-md5:	0f22124b000feb3c471096a1031d3722
 Source1:	%{name}.desktop
 Source2:	%{name}.png
 Patch0:		%{name}-libconfigure.patch
@@ -16,17 +16,19 @@ Icon:		lyx.xpm
 URL:		http://www.lyx.org/
 BuildRequires:	Aiksaurus-devel
 BuildRequires:	XFree86-devel
+BuildRequires:	aspell-devel
 BuildRequires:	autoconf
 BuildRequires:	automake
 BuildRequires:	libstdc++-devel
 BuildRequires:	qt-devel
-BuildRequires:	aspell-devel
-Prereq:		tetex
+PreReq:		tetex
 Requires:	gv
 Requires:	python-modules
 Requires:	tetex-latex
 Requires:	xdvi
 BuildRoot:	%{tmpdir}/%{name}-%{version}-root-%(id -u -n)
+
+%define		texmfdir	%{_datadir}/texmf
 
 %description
 LyX is a modern approach of writing documents with a computer which
@@ -76,15 +78,15 @@ CXXFLAGS="%{rpmcflags} -fno-exceptions"
 	--with-pspell \
 	--with-qt-includes=%{_includedir}/qt
 
-
 %{__make} all
 
 %install
 rm -rf $RPM_BUILD_ROOT
 install -d $RPM_BUILD_ROOT{%{_applnkdir}/Office/Wordprocessors,%{_pixmapsdir}} \
-	$RPM_BUILD_ROOT%{_datadir}/texmf/tex/latex/
+	$RPM_BUILD_ROOT%{texmfdir}/tex/latex
 
-%{__make} install DESTDIR=$RPM_BUILD_ROOT
+%{__make} install \
+	DESTDIR=$RPM_BUILD_ROOT
 
 install %{SOURCE1} $RPM_BUILD_ROOT%{_applnkdir}/Office/Wordprocessors
 install %{SOURCE2} $RPM_BUILD_ROOT%{_pixmapsdir}
@@ -92,7 +94,7 @@ install %{SOURCE2} $RPM_BUILD_ROOT%{_pixmapsdir}
 chmod a+rx $RPM_BUILD_ROOT%{_datadir}/lyx/configure
 
 rm -f $RPM_BUILD_ROOT%{_datadir}/lyx/{doc/LaTeXConfig.lyx,packages.lst}
-ln -sf %{_datadir}/lyx/tex $RPM_BUILD_ROOT%{_datadir}/texmf/tex/latex/lyx
+ln -sf %{_datadir}/lyx/tex $RPM_BUILD_ROOT%{texmfdir}/tex/latex/lyx
 
 %find_lang %{name}
 
@@ -101,25 +103,28 @@ rm -rf $RPM_BUILD_ROOT
 
 %post
 umask 022
-%{_bindir}/texhash
-cd %{_datadir}/lyx/
+/usr/bin/texhash
+cd %{_datadir}/lyx
 ./configure > /dev/null || :
 if [ -f lyxrc.defaults ]; then
 	cp -p lyxrc.defaults lyxrc
 fi
 
-%postun
-%{_bindir}/texhash
-
 %preun
-rm -f %{_datadir}/lyx/{lyxrc.defaults,lyxrc*}
-rm -f %{_datadir}/lyx/{doc/LaTeXConfig.lyx,packages.lst}
+if [ "$1" = "0" ]; then
+	rm -f %{_datadir}/lyx/{lyxrc.defaults,lyxrc*}
+	rm -f %{_datadir}/lyx/{doc/LaTeXConfig.lyx,packages.lst}
+fi
+
+%postun
+umask 022
+/usr/bin/texhash
 
 %files -f %{name}.lang
 %defattr(644,root,root,755)
 %doc ANNOUNCE README NEWS
 %attr(755,root,root) %{_bindir}/*
-%dir %{_datadir}/texmf/tex/latex/lyx
+%dir %{texmfdir}/tex/latex/lyx
 %attr(-, root,root) %{_datadir}/lyx
 %{_mandir}/man*/*
 %{_applnkdir}/Office/Wordprocessors/*
